+++ operator/internal/manifests/gateway_test.go
@@ -414,5 +414,216 @@ func TestBuildGateway_WithTLSProfile(t *testing.T) {
 			}
 		})
 	}
+}
 
+func TestBuildGateway_WithRulesEnabled(t *testing.T) {
+	tt := []struct {
+		desc        string
+		opts        Options
+		wantArgs    []string
+		missingArgs []string
+	}{
+		{
+			desc: "rules disabled",
+			opts: Options{
+				Name:      "abcd",
+				Namespace: "efgh",
+				Gates: configv1.FeatureGates{
+					LokiStackGateway: true,
+				},
+				Stack: lokiv1.LokiStackSpec{
+					Template: &lokiv1.LokiTemplateSpec{
+						Gateway: &lokiv1.LokiComponentSpec{
+							Replicas: rand.Int31(),
+						},
+					},
+					Tenants: &lokiv1.TenantsSpec{
+						Mode: lokiv1.Static,
+						Authorization: &lokiv1.AuthorizationSpec{
+							Roles: []lokiv1.RoleSpec{
+								{
+									Name:        "some-name",
+									Resources:   []string{"metrics"},
+									Tenants:     []string{"test-a"},
+									Permissions: []lokiv1.PermissionType{"read"},
+								},
+							},
+							RoleBindings: []lokiv1.RoleBindingsSpec{
+								{
+									Name: "test-a",
+									Subjects: []lokiv1.Subject{
+										{
+											Name: "test@example.com",
+											Kind: "user",
+										},
+									},
+									Roles: []string{"read-write"},
+								},
+							},
+						},
+					},
+				},
+			},
+			missingArgs: []string{
+				"--logs.rules.endpoint=http://abcd-ruler-http.efgh.svc.cluster.local:3100",
+				"--logs.rules.read-only=true",
+			},
+		},
+		{
+			desc: "static mode",
+			opts: Options{
+				Name:      "abcd",
+				Namespace: "efgh",
+				Gates: configv1.FeatureGates{
+					LokiStackGateway: true,
+				},
+				Stack: lokiv1.LokiStackSpec{
+					Template: &lokiv1.LokiTemplateSpec{
+						Gateway: &lokiv1.LokiComponentSpec{
+							Replicas: rand.Int31(),
+						},
+					},
+					Rules: &lokiv1.RulesSpec{
+						Enabled: true,
+					},
+					Tenants: &lokiv1.TenantsSpec{
+						Mode: lokiv1.Static,
+						Authorization: &lokiv1.AuthorizationSpec{
+							Roles: []lokiv1.RoleSpec{
+								{
+									Name:        "some-name",
+									Resources:   []string{"metrics"},
+									Tenants:     []string{"test-a"},
+									Permissions: []lokiv1.PermissionType{"read"},
+								},
+							},
+							RoleBindings: []lokiv1.RoleBindingsSpec{
+								{
+									Name: "test-a",
+									Subjects: []lokiv1.Subject{
+										{
+											Name: "test@example.com",
+											Kind: "user",
+										},
+									},
+									Roles: []string{"read-write"},
+								},
+							},
+						},
+					},
+				},
+			},
+			wantArgs: []string{
+				"--logs.rules.endpoint=http://abcd-ruler-http.efgh.svc.cluster.local:3100",
+				"--logs.rules.read-only=true",
+			},
+		},
+		{
+			desc: "dynamic mode",
+			opts: Options{
+				Name:      "abcd",
+				Namespace: "efgh",
+				Gates: configv1.FeatureGates{
+					LokiStackGateway: true,
+				},
+				Stack: lokiv1.LokiStackSpec{
+					Template: &lokiv1.LokiTemplateSpec{
+						Gateway: &lokiv1.LokiComponentSpec{
+							Replicas: rand.Int31(),
+						},
+					},
+					Rules: &lokiv1.RulesSpec{
+						Enabled: true,
+					},
+					Tenants: &lokiv1.TenantsSpec{
+						Mode: lokiv1.Dynamic,
+					},
+				},
+			},
+			wantArgs: []string{
+				"--logs.rules.endpoint=http://abcd-ruler-http.efgh.svc.cluster.local:3100",
+				"--logs.rules.read-only=true",
+			},
+		},
+		{
+			desc: "openshift-logging mode",
+			opts: Options{
+				Name:      "abcd",
+				Namespace: "efgh",
+				Gates: configv1.FeatureGates{
+					LokiStackGateway: true,
+					HTTPEncryption:   true,
+					OpenShift: configv1.OpenShiftFeatureGates{
+						ServingCertsService: true,
+					},
+				},
+				Stack: lokiv1.LokiStackSpec{
+					Template: &lokiv1.LokiTemplateSpec{
+						Gateway: &lokiv1.LokiComponentSpec{
+							Replicas: rand.Int31(),
+						},
+					},
+					Rules: &lokiv1.RulesSpec{
+						Enabled: true,
+					},
+					Tenants: &lokiv1.TenantsSpec{
+						Mode: lokiv1.OpenshiftLogging,
+					},
+				},
+			},
+			wantArgs: []string{
+				"--logs.rules.endpoint=https://abcd-ruler-http.efgh.svc.cluster.local:3100",
+				"--logs.rules.read-only=true",
+			},
+		},
+		{
+			desc: "openshift-network mode",
+			opts: Options{
+				Name:      "abcd",
+				Namespace: "efgh",
+				Gates: configv1.FeatureGates{
+					LokiStackGateway: true,
+					HTTPEncryption:   true,
+					OpenShift: configv1.OpenShiftFeatureGates{
+						ServingCertsService: true,
+					},
+				},
+				Stack: lokiv1.LokiStackSpec{
+					Template: &lokiv1.LokiTemplateSpec{
+						Gateway: &lokiv1.LokiComponentSpec{
+							Replicas: rand.Int31(),
+						},
+					},
+					Rules: &lokiv1.RulesSpec{
+						Enabled: true,
+					},
+					Tenants: &lokiv1.TenantsSpec{
+						Mode: lokiv1.OpenshiftLogging,
+					},
+				},
+			},
+			wantArgs: []string{
+				"--logs.rules.endpoint=https://abcd-ruler-http.efgh.svc.cluster.local:3100",
+				"--logs.rules.read-only=true",
+			},
+		},
+	}
+	for _, tc := range tt {
+		tc := tc
+		t.Run(tc.desc, func(t *testing.T) {
+			t.Parallel()
+			objs, err := BuildGateway(tc.opts)
+			require.NoError(t, err)
+
+			d, ok := objs[1].(*appsv1.Deployment)
+			require.True(t, ok)
+
+			for _, arg := range tc.wantArgs {
+				require.Contains(t, d.Spec.Template.Spec.Containers[0].Args, arg)
+			}
+			for _, arg := range tc.missingArgs {
+				require.NotContains(t, d.Spec.Template.Spec.Containers[0].Args, arg)
+			}
+		})
+	}
 }