+++ pkg/loki/config_wrapper_test.go
@@ -3,7 +3,6 @@ package loki
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"net/url"
 	"os"
 	"reflect"
@@ -39,7 +38,7 @@ func configWrapperFromYAML(t *testing.T, configFileString string, args []string)
 	config := ConfigWrapper{}
 	fs := flag.NewFlagSet(t.Name(), flag.PanicOnError)
 
-	file, err := ioutil.TempFile("", "config.yaml")
+	file, err := os.CreateTemp("", "config.yaml")
 	defer func() {
 		os.Remove(file.Name())
 	}()
@@ -987,7 +986,7 @@ query_range:
 
 func TestDefaultUnmarshal(t *testing.T) {
 	t.Run("with a minimal config file and no command line args, defaults are use", func(t *testing.T) {
-		file, err := ioutil.TempFile("", "config.yaml")
+		file, err := os.CreateTemp("", "config.yaml")
 		defer func() {
 			os.Remove(file.Name())
 		}()