+++ pkg/storage/stores/indexshipper/compactor/deletion/delete_requests_client.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"sync"
@@ -167,7 +166,7 @@ func (c *deleteRequestsClient) getRequestsFromServer(ctx context.Context, userID
 		return nil, err
 	}
 	defer func() {
-		_, _ = io.Copy(ioutil.Discard, resp.Body)
+		_, _ = io.Copy(io.Discard, resp.Body)
 		_ = resp.Body.Close()
 	}()