+++ pkg/storage/stores/indexshipper/downloads/table_test.go
@@ -2,7 +2,7 @@ package downloads
 
 import (
 	"context"
-	"io/ioutil"
+	"io"
 	"os"
 	"path/filepath"
 	"sort"
@@ -301,8 +301,8 @@ func TestTable_Sync(t *testing.T) {
 	newDB := "new"
 
 	require.NoError(t, os.MkdirAll(tablePathInStorage, 0755))
-	require.NoError(t, ioutil.WriteFile(filepath.Join(tablePathInStorage, deleteDB), []byte(deleteDB), 0755))
-	require.NoError(t, ioutil.WriteFile(filepath.Join(tablePathInStorage, noUpdatesDB), []byte(noUpdatesDB), 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(tablePathInStorage, deleteDB), []byte(deleteDB), 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(tablePathInStorage, noUpdatesDB), []byte(noUpdatesDB), 0755))
 
 	// create table instance
 	table, stopFunc := buildTestTable(t, tempDir)
@@ -326,7 +326,7 @@ func TestTable_Sync(t *testing.T) {
 
 	// remove deleteDB and add the newDB
 	require.NoError(t, os.Remove(filepath.Join(tablePathInStorage, deleteDB)))
-	require.NoError(t, ioutil.WriteFile(filepath.Join(tablePathInStorage, newDB), []byte(newDB), 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(tablePathInStorage, newDB), []byte(newDB), 0755))
 
 	// sync the table
 	table.storageClient.RefreshIndexListCache(context.Background())
@@ -347,13 +347,13 @@ func TestTable_Sync(t *testing.T) {
 		noUpdatesDB: {},
 		newDB:       {},
 	}
-	filesInfo, err := ioutil.ReadDir(tablePathInStorage)
+	dirEntries, err := os.ReadDir(tablePathInStorage)
 	require.NoError(t, err)
 	require.Len(t, table.indexSets[""].(*indexSet).index, len(expectedFilesInDir))
 
-	for _, fileInfo := range filesInfo {
-		require.False(t, fileInfo.IsDir())
-		_, ok := expectedFilesInDir[fileInfo.Name()]
+	for _, entry := range dirEntries {
+		require.False(t, entry.IsDir())
+		_, ok := expectedFilesInDir[entry.Name()]
 		require.True(t, ok)
 	}
 
@@ -361,12 +361,12 @@ func TestTable_Sync(t *testing.T) {
 
 	// first, let us add a new file and refresh the index list cache
 	oneMoreDB := "one-more-db"
-	require.NoError(t, ioutil.WriteFile(filepath.Join(tablePathInStorage, oneMoreDB), []byte(oneMoreDB), 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(tablePathInStorage, oneMoreDB), []byte(oneMoreDB), 0755))
 	table.storageClient.RefreshIndexListCache(context.Background())
 
 	// now, without syncing the table, let us compact the index in storage
 	compactedDBName := "compacted-db"
-	require.NoError(t, ioutil.WriteFile(filepath.Join(tablePathInStorage, compactedDBName), []byte(compactedDBName), 0755))
+	require.NoError(t, os.WriteFile(filepath.Join(tablePathInStorage, compactedDBName), []byte(compactedDBName), 0755))
 	require.NoError(t, os.Remove(filepath.Join(tablePathInStorage, noUpdatesDB)))
 	require.NoError(t, os.Remove(filepath.Join(tablePathInStorage, newDB)))
 	require.NoError(t, os.Remove(filepath.Join(tablePathInStorage, oneMoreDB)))
@@ -461,7 +461,7 @@ func verifyIndexForEach(t *testing.T, expectedIndexes []string, forEachFunc func
 		require.NoError(t, err)
 
 		// read the contents of the index.
-		buf, err := ioutil.ReadAll(readSeeker)
+		buf, err := io.ReadAll(readSeeker)
 		require.NoError(t, err)
 
 		// see if it matches the name of the file